fix: update all allow-list references to allowlist
